CORE-1298 Move NotificationsMenu content to tab
