+++ src/components/notifications/NotificationsTab.js
@@ -1,0 +1,236 @@
+import React from "react";
+
+import { build } from "@cyverse-de/ui-lib";
+import {
+    Button,
+    Divider,
+    IconButton,
+    ListItem,
+    ListItemText,
+    makeStyles,
+    Typography,
+    useMediaQuery,
+    useTheme,
+} from "@material-ui/core";
+import DoneAllIcon from "@material-ui/icons/DoneAll";
+import OpenInNewIcon from "@material-ui/icons/OpenInNew";
+import { Skeleton } from "@material-ui/lab";
+import classnames from "classnames";
+import { formatDistance, fromUnixTime } from "date-fns";
+import Link from "next/link";
+
+import NavigationConstants from "common/NavigationConstants";
+import ErrorTypographyWithDialog from "components/utils/error/ErrorTypographyWithDialog";
+import { useTranslation } from "i18n";
+import ids from "./ids";
+import Message from "./Message";
+import styles from "./styles";
+
+const useStyles = makeStyles(styles);
+
+/*
+ * Takes in a notification object and returns the time
+ * stamp of the notification in 'pretty format'
+ */
+function getTimeStamp(timestamp) {
+    if (timestamp) {
+        // slicing because time has extra zeroes in the unix string
+        const d = fromUnixTime(timestamp.slice(0, -3));
+        return formatDistance(d, new Date());
+    }
+}
+
+const NotificationsListingButton = React.forwardRef((props, ref) => {
+    const { isMobile, handleClose, href, onClick } = props;
+    const { t } = useTranslation("common");
+    const buttonId = build(
+        ids.BASE_DEBUG_ID,
+        ids.NOTIFICATIONS_MENU,
+        ids.VIEW_ALL_NOTIFICATIONS
+    );
+
+    return isMobile ? (
+        <IconButton
+            className={useStyles().viewAll}
+            id={buttonId}
+            ref={ref}
+            href={href}
+            onClick={(event) => {
+                onClick(event);
+                handleClose();
+            }}
+        >
+            <OpenInNewIcon size="small" />
+        </IconButton>
+    ) : (
+        <Button
+            id={buttonId}
+            color="primary"
+            startIcon={<OpenInNewIcon size="small" />}
+            ref={ref}
+            href={href}
+            onClick={(event) => {
+                onClick(event);
+                handleClose();
+            }}
+        >
+            {t("viewAllNotifications")}
+        </Button>
+    );
+});
+
+function NotificationsListingLink(props) {
+    const href = `/${NavigationConstants.NOTIFICATIONS}`;
+
+    return (
+        <Link href={href} as={href} passHref>
+            <NotificationsListingButton {...props} />
+        </Link>
+    );
+}
+
+function NotificationsTab(props) {
+    const {
+        isFetching,
+        notifications,
+        handleClose,
+        handleMarkAllAsSeenClick,
+        errorObject,
+    } = props;
+
+    const classes = useStyles();
+    const theme = useTheme();
+    const isMobile = useMediaQuery(theme.breakpoints.down("xs"));
+    const { t } = useTranslation("common");
+
+    return (
+        <>
+            {isMobile && [
+                <NotificationsListingLink
+                    key={ids.VIEW_ALL_NOTIFICATIONS}
+                    id={build(
+                        ids.BASE_DEBUG_ID,
+                        ids.NOTIFICATIONS_MENU,
+                        ids.VIEW_ALL_NOTIFICATIONS
+                    )}
+                    handleClose={handleClose}
+                    isMobile={isMobile}
+                />,
+                <IconButton
+                    key={ids.MARK_ALL_READ}
+                    className={classes.markSeen}
+                    onClick={handleMarkAllAsSeenClick}
+                    id={build(
+                        ids.BASE_DEBUG_ID,
+                        ids.NOTIFICATIONS_MENU,
+                        ids.MARK_ALL_READ
+                    )}
+                >
+                    <DoneAllIcon size="small" />
+                </IconButton>,
+            ]}
+            {isFetching && (
+                <Skeleton variant="rect" height={400} animation="wave" />
+            )}
+            {!isFetching && errorObject !== null && (
+                <ListItem>
+                    <ErrorTypographyWithDialog
+                        errorMessage={t("notificationError", errorObject)}
+                    />
+                </ListItem>
+            )}
+
+            {!isFetching &&
+                errorObject === null &&
+                (notifications === null || notifications.length === 0) && (
+                    <ListItem>
+                        <Typography variant="body2">
+                            {t("noNotifications")}
+                        </Typography>
+                    </ListItem>
+                )}
+
+            {!isFetching &&
+                notifications.length > 0 &&
+                notifications.map((n, index) => (
+                    <ListItem
+                        onClick={handleClose}
+                        id={build(
+                            ids.BASE_DEBUG_ID,
+                            ids.NOTIFICATIONS_MENU,
+                            index
+                        )}
+                        key={n.message.id}
+                        className={
+                            !n.seen
+                                ? classnames(
+                                      classes.notification,
+                                      classes.unSeenNotificationBackground
+                                  )
+                                : classes.notification
+                        }
+                        dense={true}
+                        divider={true}
+                    >
+                        <ListItemText
+                            primary={
+                                <Message
+                                    baseId={build(
+                                        ids.BASE_DEBUG_ID,
+                                        ids.NOTIFICATIONS_MENU,
+                                        n?.message.id
+                                    )}
+                                    notification={n}
+                                />
+                            }
+                            secondary={
+                                <Typography
+                                    id={build(
+                                        ids.BASE_DEBUG_ID,
+                                        ids.NOTIFICATIONS_MENU,
+                                        ids.TIME_STAMP
+                                    )}
+                                    variant="caption"
+                                >
+                                    {t("timestamp", {
+                                        timestamp: getTimeStamp(
+                                            n.message?.timestamp
+                                        ),
+                                    })}
+                                </Typography>
+                            }
+                        />
+                    </ListItem>
+                ))}
+
+            {!isMobile && [
+                <Divider light key="divider" />,
+                <NotificationsListingLink
+                    key={ids.VIEW_ALL_NOTIFICATIONS}
+                    id={build(
+                        ids.BASE_DEBUG_ID,
+                        ids.NOTIFICATIONS_MENU,
+                        ids.VIEW_ALL_NOTIFICATIONS
+                    )}
+                    handleClose={handleClose}
+                    isMobile={isMobile}
+                />,
+                <Button
+                    key={ids.MARK_ALL_READ}
+                    id={build(
+                        ids.BASE_DEBUG_ID,
+                        ids.NOTIFICATIONS_MENU,
+                        ids.MARK_ALL_READ
+                    )}
+                    color="primary"
+                    onClick={handleMarkAllAsSeenClick}
+                    startIcon={<DoneAllIcon size="small" />}
+                >
+                    {t("markAsRead")}
+                </Button>,
+            ]}
+        </>
+    );
+}
+
+export default NotificationsTab;
